requirements.txt
Django~=2.2.4
django-tinymce4-lite
requests